Experimental fix for Astro JS
These features ensure that typescript is required and that our astro.js
files are read properly. NOTE As of this time, the process
intermittently works 100%. When first opening an astro file with this
config, in some cases it works immediately in other cases it gives the
typescript cannot be located error.